Fixed mouse trap escape issue
While the mask can make a shot, its x and z-positions remain constrained. As a result, they will not teleport out of the mouse traps, and in places where the masks may get stuck in awkward angles, the masks will at least remain around the same location and at about the same rotation.