Fixed roomba sucking error
The roomba will now absorb masks even when they're not moving, since the roombas now have rigidbodies that keep track of collisions for static objects. Edits for the masks need to be made to allow the masks to either avoid absorption when aiming, or wait until they have touched a roomba after entering some trigger indicating the roomba's patrol path.